backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "google-cloud-texttospeech>=2.36.0",
    "google-genai>=1.73.1",
    "pydub>=0.25.1",
    "pyopenssl>=26.0.0",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "python-pptx>=1.0.2",
    "uvicorn>=0.44.0",
]
